Add files via upload
The code to use for the following car to track the lead car by going forward if an object is detected within a certain range in front. It's ultrasonic sensor will turn left and right and turn in the direction in which it senses an object if the front object moves out of view.